requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
colorama==0.4.6
Werkzeug==2.3.7
